Moved gmcp functions into Lua directory and renamed scripts.lua to oracle.lua.
